COMMERCE-4433 Refactor CommerceContext* -> BaseCommerceContext* so classes can't be extended
