+++ modules/apps/commerce/commerce-test/src/testIntegration/java/com/liferay/commerce/internal/order/test/CommerceOrderTest.java
@@ -25,6 +25,7 @@ import com.liferay.commerce.account.util.CommerceAccountHelper;
 import com.liferay.commerce.constants.CommerceAddressConstants;
 import com.liferay.commerce.constants.CommerceConstants;
 import com.liferay.commerce.constants.CommerceOrderConstants;
+import com.liferay.commerce.context.CommerceContext;
 import com.liferay.commerce.currency.model.CommerceCurrency;
 import com.liferay.commerce.currency.test.util.CommerceCurrencyTestUtil;
 import com.liferay.commerce.exception.CommerceOrderAccountLimitException;
@@ -37,6 +38,9 @@ import com.liferay.commerce.product.service.CommerceChannelLocalService;
 import com.liferay.commerce.service.CommerceAddressLocalService;
 import com.liferay.commerce.service.CommerceOrderLocalService;
 import com.liferay.commerce.service.CommerceOrderService;
+import com.liferay.commerce.service.CommerceRegionLocalService;
+import com.liferay.commerce.test.util.CommerceTestUtil;
+import com.liferay.commerce.test.util.TestCommerceContext;
 import com.liferay.petra.lang.CentralizedThreadLocal;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
@@ -140,6 +144,25 @@ public class CommerceOrderTest {
 			_commerceChannel.getGroupId());
 
 		CentralizedThreadLocal.clearShortLivedThreadLocals();
+	}
+
+	@Test
+	public void testCommerceOrderCurrency() throws Exception {
+		CommerceCurrency commerceCurrency2 = CommerceCurrencyTestUtil.addCommerceCurrency(
+			_group.getCompanyId());
+
+		CommerceOrder commerceOrder = CommerceTestUtil.addB2CCommerceOrder(
+			_user.getUserId(), _group.getGroupId(), commerceCurrency2);
+
+		Assert.assertEquals(
+			commerceCurrency2, commerceOrder.getCommerceCurrency());
+
+		CommerceCurrency orderCommerceCurrency =
+			commerceOrder.getCommerceCurrency();
+
+		Assert.assertNotEquals(
+			orderCommerceCurrency.getCode(),
+			_commerceChannel.getCommerceCurrencyCode());
 	}
 
 	@Test
